ARM: tegra: fix warning
This fixes the following warning by casting the integer to the correct
size before casting it to a pointer:

arch/arm/mach-tegra/tegra210/nvtboot.c: In function 'fdt_copy_get_blob_src_default':
arch/arm/mach-tegra/tegra210/nvtboot.c:469:9: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
  return (void *)nvtboot_boot_arg.kernel_dtb_load_addr;
         ^

Bug 1831386

Reviewed-on: http://git-master/r/1244003
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ arch/arm/mach-tegra/tegra210/nvtboot.c
@@ -466,5 +466,5 @@ void board_cleanup_before_linux(void)
 
 void *fdt_copy_get_blob_src_default(void)
 {
-	return (void *)nvtboot_boot_arg.kernel_dtb_load_addr;
+	return (void *)(unsigned long)nvtboot_boot_arg.kernel_dtb_load_addr;
 }
